Crate secure_serialisation

Secure Serialisation

Given a remote nacl box PublicKey this lib will securely serialise and encrypt messages destined for that node. These will use authenticated encryption.

Authenticated encryption

Using public-key authenticated encryption, Bob can encrypt a confidential message specificall for Alice, using Alice's public key.+

Using Bob's public key, Alice can verify that the encrypted message was actually created by Bob and was not tampered with, before eventually decrypting it. Alice only needs Bob's public key, the nonce and the ciphertext. Bob should never ever share his secret key, even with Alice. And in order to send messages to Alice, Bob only needs Alice's public key. Alice should never ever share her secret key either, even with Bob. Alice can reply to Bob using the same system, without having to generate a distinct key pair. The nonce doesn't have to be confidential, but it should be used with just one invocation of crypto_box_open_easy() for a particular pair of public and secret keys. One easy way to generate a nonce is to use randombytes_buf(), considering the size of nonces the risk of any random collisions is negligible. For some applications, if you wish to use nonces to detect missing messages or to ignore replayed messages, it is also ok to use a simple incrementing counter as a nonce. In this crate we use a random nonce wrapped into the message. When doing so you must ensure that the same value can never be re-used (for example you may have multiple threads or even hosts generating messages using the same key pairs). This system provides mutual authentication. However, a typical use case is to secure communications between a server, whose public key is known in advance, and clients connecting anonymously.

This implementation will encrypt data with a nonce and then serialise the payload. The nonce is then prepended to the beginning of the message and pulled off first at the remote end. Thsi provides a clean secure mechnism fro sending data between entities who have session based keypairs. It SHOULD NOT be used for permenent keys.

Enums

SecureSerialisationError
Error types, hopefully sodiumoxide eventually defines errors properly, otherwise this makes little sense really

Functions

deserialise
Parse a data type from an ecnoded message, sucess ensures teh message was from the holder of the private_key related to the public_key we know of the recipient

serialise
Prepare an ecodable data element for transmission to another process whose public_key we know.